+++ calcdialog/src/main/java/com/nmaltais/calcdialog/CalcEraseButton.java
@@ -29,6 +29,7 @@ import android.os.Handler;
 import android.support.annotation.Nullable;
 import android.support.v7.widget.AppCompatImageView;
 import android.util.AttributeSet;
+import android.view.HapticFeedbackConstants;
 import android.view.MotionEvent;
 
 /**
@@ -38,15 +39,18 @@ import android.view.MotionEvent;
  *                      Default value is 750ms. Use -1 for no quick erase and 0 for no delay
  * - eraseBtnHoldSpeed: Time after which an erase event is triggered in quick erase mode (in ms)
  *                      Default value is 100ms
+ * - eraseAllOnHold: If true, holding button will trigger an erase all event instead of quick
+ *                   erase mode if false. By default this is false.
  */
-class EraseButton extends AppCompatImageView {
+class CalcEraseButton extends AppCompatImageView {
 
-    private static final String TAG = EraseButton.class.getSimpleName();
+    private static final String TAG = CalcEraseButton.class.getSimpleName();
 
     private static final int NO_HOLD_ERASE = -1;
 
     private int eraseHoldDelay;
     private int eraseHoldSpeed;
+    private boolean eraseAllOnHold;
 
     private final Handler eraseHandler;
     private final Runnable eraseRunnable;
@@ -54,21 +58,22 @@ class EraseButton extends AppCompatImageView {
 
     private @Nullable EraseListener listener;
 
-    public EraseButton(Context context) {
+    public CalcEraseButton(Context context) {
         this(context, null, 0);
     }
 
-    public EraseButton(Context context, @Nullable AttributeSet attrs) {
+    public CalcEraseButton(Context context, @Nullable AttributeSet attrs) {
         this(context, attrs, 0);
     }
 
-    public EraseButton(Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
+    public CalcEraseButton(Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
 
         // Get speed attributes
-        TypedArray ta = getContext().obtainStyledAttributes(attrs, R.styleable.EraseButton);
-        eraseHoldDelay = ta.getInt(R.styleable.EraseButton_eraseBtnHoldDelay, 750);
-        eraseHoldSpeed = ta.getInt(R.styleable.EraseButton_eraseBtnHoldSpeed, 100);
+        TypedArray ta = getContext().obtainStyledAttributes(attrs, R.styleable.CalcEraseButton);
+        eraseHoldDelay = ta.getInt(R.styleable.CalcEraseButton_calcEraseBtnHoldDelay, 750);
+        eraseHoldSpeed = ta.getInt(R.styleable.CalcEraseButton_calcEraseBtnHoldSpeed, 100);
+        eraseAllOnHold = ta.getBoolean(R.styleable.CalcEraseButton_calcEraseAllOnHold, false);
         ta.recycle();
 
         eraseHandler = new Handler();
@@ -76,8 +81,12 @@ class EraseButton extends AppCompatImageView {
             @Override
             public void run() {
                 if (listener != null && clickingDown) {
-                    listener.onErase();
-                    eraseHandler.postDelayed(eraseRunnable, eraseHoldSpeed);
+                    if (eraseAllOnHold) {
+                        listener.onEraseAll();
+                    } else {
+                        listener.onErase();
+                        eraseHandler.postDelayed(eraseRunnable, eraseHoldSpeed);
+                    }
                 }
             }
         };
@@ -85,6 +94,7 @@ class EraseButton extends AppCompatImageView {
 
     public interface EraseListener {
         void onErase();
+        void onEraseAll();
     }
 
     public void setOnEraseListener(@Nullable EraseListener listener) {
@@ -109,7 +119,16 @@ class EraseButton extends AppCompatImageView {
             if (listener != null) {
                 if (eraseHoldDelay != NO_HOLD_ERASE) {
                     eraseHandler.postDelayed(eraseRunnable, eraseHoldDelay);
+                    eraseHandler.postDelayed(new Runnable() {
+                        @Override
+                        public void run() {
+                            if (clickingDown) {
+                                performHapticFeedback(HapticFeedbackConstants.LONG_PRESS);
+                            }
+                        }
+                    }, eraseHoldDelay);
                 }
+
                 if (eraseHoldDelay != 0) {
                     listener.onErase();
                 }